Move more helpers to helpers
